+++ user/repository/postgres/postgres_repository.go
@@ -1,8 +1,10 @@
-package user
+package postgres
 
 import (
 	"context"
 	"database/sql"
+
+	"github.com/RianNegreiros/vigilate/domain"
 )
 
 type DBTX interface {
@@ -16,29 +18,29 @@ type repository struct {
 	db DBTX
 }
 
-func NewRepository(db DBTX) Repository {
+func NewRepository(db DBTX) domain.UserRepository {
 	return &repository{db: db}
 }
 
-func (r *repository) CreateUser(ctx context.Context, user *User) (*User, error) {
+func (r *repository) CreateUser(ctx context.Context, user *domain.User) (*domain.User, error) {
 	var lastInsertId int
 	query := "INSERT INTO users(username, password, email) VALUES ($1, $2, $3) returning id"
 
 	err := r.db.QueryRowContext(ctx, query, user.Username, user.Password, user.Email).Scan(&lastInsertId)
 	if err != nil {
-		return &User{}, err
+		return &domain.User{}, err
 	}
 
 	user.ID = int64(lastInsertId)
 	return user, nil
 }
 
-func (r *repository) GetUserByEmail(ctx context.Context, email string) (*User, error) {
-	u := User{}
+func (r *repository) GetUserByEmail(ctx context.Context, email string) (*domain.User, error) {
+	u := domain.User{}
 	query := "SELECT id, email, username, password FROM users WHERE email = $1"
 	err := r.db.QueryRowContext(ctx, query, email).Scan(&u.ID, &u.Email, &u.Username, &u.Password)
 	if err != nil {
-		return &User{}, err
+		return &domain.User{}, err
 	}
 
 	return &u, nil
